feat: 에러페이지 제작 완료
