+++ Tripeer/src/App.tsx
@@ -85,10 +85,6 @@ const router = createBrowserRouter([
         element: <p>마이페이지</p>,
         loader: protectRouter(),
       },
-      {
-        path: '/error',
-        element: <ErrorPage />,
-      },
     ],
   },
   {
